added new state management for Pomodoro settings and organised file system
