+++ src/components/BottomSheet/CreateSessionModal.jsx
@@ -8,7 +8,7 @@ import DurationModal from './DurationModal';
 import { router } from 'expo-router';
 import { faTag, faCaretDown } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
-import TaskSelector from '@/components/TaskSelector';
+import TaskSelector from '@/components/BottomSheet/TaskSelector';
 import { CustomSvg } from '@/components/CustomSvg';
 import CustomButton from '@/components/CustomButton';
 import Animated, { FadeIn, FadeInDown, FadeInUp, FadeOut } from 'react-native-reanimated';
@@ -16,7 +16,7 @@ import { BlurView } from '@react-native-community/blur';
 import SegmentadControl from '@/components/SegmentadControl';
 
 import TimerBlock from '@/components/TimerBlock';
-import Pomodoro from '@/components/Pomodoro';
+import Pomodoro from '@/components/Timer/Pomodoro';
 const { width: SCREEN_WIDTH } = Dimensions.get('window');
 
 export const CreateSessionModal = ({ bottomSheetModalRef }) => {
